separated helper functions into their own files and made a master index for help functions to pull from
