+++ src/utils/helper-functions/index.js
@@ -1,0 +1,5 @@
+import { mapSetUp } from './mapSetUp';
+import { popUpSetUp } from './popUpSetUp';
+import { markerSetUp } from './markerSetUp';
+
+export { mapSetUp, popUpSetUp, markerSetUp };
